LUN-4: Move and rename tests
