+++ package.json
@@ -8,6 +8,7 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest --silent --verbose",
+    "test:update": "yarn run test -u",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "lint:fix": "eslint . --fix --ext .js,.jsx,.ts,.tsx",
     "lint:ci": "yarn run eslint --quiet --format junit -o reports/lint/junit-lint.xml",
