+++ src/components/__tests__/Actions.spec.tsx
@@ -1,10 +1,10 @@
 import 'react-native';
 import React from 'react';
-import Action from '../../src/components/Actions';
+import Action from '../Actions';
 import {shallow} from 'enzyme';
 import toJson from 'enzyme-to-json';
-import {IActionsProps} from '../../src/interfaces/exercise';
-import {styles} from '../../src/components/Actions';
+import {IActionsProps} from '../../interfaces/exercise';
+import {styles} from '../Actions';
 
 describe('Components', () => {
   describe('Actions', () => {
