+++ src/components/__tests__/Header.spec.tsx
@@ -1,9 +1,9 @@
 import 'react-native';
 import React from 'react';
-import Header from '../../src/components/Header';
+import Header from '../Header';
 import {shallow} from 'enzyme';
 import toJson from 'enzyme-to-json';
-import {IHeaderProps} from '../../src/interfaces/header';
+import {IHeaderProps} from '../../interfaces/header';
 
 describe('Components', () => {
   describe('Header', () => {
