+++ src/components/__tests__/Title.spec.tsx
@@ -1,10 +1,10 @@
 import 'react-native';
 import React from 'react';
-import Title from '../../src/components/Title';
+import Title from '../Title';
 import {shallow} from 'enzyme';
 import toJson from 'enzyme-to-json';
 import {Text} from 'react-native';
-import {ITitleProps} from '../../src/interfaces/title';
+import {ITitleProps} from '../../interfaces/title';
 
 describe('Components', () => {
   describe('Title', () => {
